Move Executor and C api references into CApiState
Api and its subtypes just have a reference to CApiState which has
all the real references.

CApiState is the pointer passed to update_callback().

We still leave Executor as a pointer on CApiState because we want
to have interior mutability on the Executor. And while RefCell
would let us do so with safe Rust, we'd have a bad time storing
references to the Executor from Wakers then.