Move frame number out of executor, introduce FrameWaiter
Instead of having an async function that waits for a global state, have an
object which can be queried for the next frame with .next().

A select() would return when the first future is ready, but a
Future watching a global state is not very friendly for this. If
the Future isn't first in the select(), there's no way to wait
again for the same frame.  Since Futures aren't visible to async
code, we make a FrameWaiter type that is visible and has a
consistent state so that even if a request is dropped, the next
request will wait for the same thing.