Add System::get_current_time() which calls getCurrentTimeMilliseconds()
Introduce TimeTicks and TimeDelta types. Internally TimeTicks is a
u32 holding milliseconds since that is the precision that Playdate
provides through getCurrentTimeMilliseconds().

We don't expose the internal value directly, but you can ask for the
value represented as hours, minutes, seconds, or milliseconds.
They each truncate any non-whole parts of the time in that
representation.

If Playdate one day supported a higher resolution, then we could
make a backward compatible change and add a function to
provide microseconds.

TimeDelta, like TimeTicks, uses a 32bit integer. But it uses a signed
integer since deltas can be negative.

Add System::get_current_time() which returns TimeTicks. This is
critically different from Playdate's getSecondsSinceEpoch which
will return an actual world time, and which would change when
the device's timezone or clock was changed.