Expose button presses and releases on Inputs
To do so reliably, we need to track the button states from the
previous frame.  We hold this on CApiState. And in order to avoid
mutating the CApiState at the "end" of a frame, we keep both the
current and previous frames' button states, and update them
when the FrameWatcher returns Ready(Inputs).

The Inputs infers button events by combinings the previous frame
state, current frame state, and whether a push/release happened
between frames. For example

If A was released and is now pressed, we know the sequence of (observable) events was either:
- Push, or
- Push Release Push
And we can differentiate between these two options by checking
if a release happened in between frames.

On Inputs we keep this complexity hidden as best as possible,
while steering people toward a well-lit path.

We expose the set of events that happened for each button as an
iterator for each button. And we also provide a combination of all
buttons together on a single iterator via all_events(). The ordering
between buttons is not exposed by Playdate, so we just pick an
arbitrary order.

We don't expose the raw Playdate state for if a button was
pushed/released between frames. But we do expose the current
button state so apps don't have to track a bool themselves if they
just want a current state.