Fix mut-ness and types in graphics mod
Also:

Grab the pixels off the LCDBitmap directly, the borrow checker ensures they don't outlive the bitmap. The LCDBitmapData is just a copy/cloneable struct of ints now.

Pass &str to most functions that want a string, to be clear they don't need to take ownership and that they should receive a string. Not random types that can become strings. https://users.rust-lang.org/t/when-to-use-asref-t-vs-t/9312/4