Add Sequence, SequenceTrack, and Control for MIDI files
A few TODOs in there for missing things, but mostly there.

Tracks are set on a Sequence, but the Sequence doesn't own them.
However you can load a Sequence from a file, in which case it
would own them.

We could consider making those different kinds of Sequences I
suppose. For now, we just require tracks to outlive the sequence,
and that means setting them all when you construct the sequence.